+++ changelog.md
@@ -7,6 +7,7 @@
 
 _(most recent changes are listed on top):_
 - Initial release
+- Add Tag Filter feature from [Taglist](https://github.com/TimoStahl/kanboard_plugin_taglist) plugin @timostahl
 - Added `en_GB` translations